+++ src/app/features/directory/actions.ts
@@ -1,31 +1,8 @@
 "use server";
 
-import { getApolloClient, getApolloClientAuthed } from "~/lib/apollo-client";
+import { getApolloClientAuthed } from "~/lib/apollo-client";
 import { gql } from "@apollo/client";
-import { getNFTContracts } from "~/lib/getNFTContracts";
-import {
-  Creator,
-  CreatorWithNFTData,
-  Give,
-  GroupedGives,
-  SortedGiveGroup,
-} from "~/app/features/directory/types";
-import { resolveBasenameOrAddress } from "~/app/hooks/useBasenameResolver";
-import { kv } from "@vercel/kv";
-import {
-  CIRCLE_ID,
-  ENTRANCE,
-  CREATORS_CACHE_KEY,
-  REVALIDATION_LOCK_KEY,
-  CACHE_DURATION,
-  LOCK_DURATION,
-  REVALIDATION_WINDOW,
-} from "./constants";
-import { debugLog } from "~/lib/constants";
-
-export interface CachedData {
-  data: CreatorWithNFTData[];
-}
+import { CIRCLE_ID, ENTRANCE } from "./constants";
 
 /**
  * Checks if an address is a member of the hardcoded circle
@@ -109,446 +86,5 @@ export async function joinDirectory(
   } catch (error) {
     console.error("Error joining directory:", error);
     return false;
-  }
-}
-
-/**
- * Fetches a single creator from the directory by their Ethereum address
- * @param address Ethereum address of the creator to fetch
- * @returns Promise<CreatorWithNFTData | null> The creator with their NFT data or null if not found
- */
-export async function getCreator(
-  address: string,
-): Promise<CreatorWithNFTData | null> {
-  try {
-    const { data } = await getApolloClientAuthed().query({
-      query: gql`
-        query CreatorsDirGetSingleCreator(
-          $circleId: bigint!
-          $address: String!
-        ) {
-          users(
-            where: {
-              circle_id: { _eq: $circleId }
-              profile: { address: { _ilike: $address } }
-            }
-          ) {
-            id
-            profile {
-              id
-              address
-              name
-              avatar
-              description
-              farcaster_account {
-                username
-              }
-            }
-          }
-        }
-      `,
-      variables: {
-        circleId: CIRCLE_ID,
-        address: address,
-      },
-    });
-
-    // If no creator found with this address
-    if (!data.users.length) {
-      return null;
-    }
-
-    // Transform the data to a more convenient format (using the first result)
-    const user = data.users[0];
-    const creator: Creator = {
-      id: user.id,
-      address: user.profile?.address || "",
-      name: user.profile?.name || "",
-      avatar: user.profile?.avatar
-        ? user.profile.avatar.startsWith("http")
-          ? user.profile.avatar
-          : `https://coordinape-prod.s3.amazonaws.com/${user.profile.avatar}`
-        : "",
-      description: user.profile?.description || "",
-      farcasterUsername: user.profile?.farcaster_account?.username || "",
-    };
-
-    try {
-      // Get NFT contracts
-      const contracts = await getNFTContracts(creator.address, "BASE_MAINNET", {
-        excludeNoImage: true,
-      });
-      console.log("getNFTContracts", JSON.stringify(contracts, null, 2));
-
-      // Resolve basename
-      const resolution = await resolveBasenameOrAddress(creator.address);
-
-      // Transform the resolution to match the BasenameResolution interface
-      const formattedResolution = resolution
-        ? {
-            basename: resolution.basename,
-            address: resolution.address,
-            resolved: !!resolution.basename,
-            textRecords: resolution.textRecords,
-          }
-        : null;
-
-      return {
-        ...creator,
-        resolution: formattedResolution,
-        nftData: {
-          collections: contracts.map((contract) => ({
-            id: contract.contractAddress,
-            name: contract.name,
-            description: contract.description,
-            imageUrl: contract.imageUrl,
-            bannerImageUrl: contract.bannerImageUrl,
-            openseaUrl: `https://opensea.io/assets/base/${contract.contractAddress}`,
-            projectUrl: contract.projectUrl || "",
-            contractAddress: contract.contractAddress,
-          })),
-        },
-      };
-    } catch (error) {
-      console.error(`Failed to fetch data for ${creator.address}:`, error);
-      // Return creator without NFT data if there's an error
-      return {
-        ...creator,
-        resolution: null,
-        nftData: { collections: [] },
-      };
-    }
-  } catch (error) {
-    console.error(`Error fetching creator with address ${address}:`, error);
-    return null;
-  }
-}
-
-/**
- * Internal function to refresh creators data and update cache
- * @returns Promise<CreatorWithNFTData[]> Array of creators with their NFT data
- */
-async function refreshCreatorsData(): Promise<CreatorWithNFTData[]> {
-  const freshData = await getCreatorsFromAPI();
-  await kv.set(
-    CREATORS_CACHE_KEY,
-    {
-      data: freshData,
-    },
-    { ex: CACHE_DURATION },
-  );
-  return freshData;
-}
-
-/**
- * Fetches all creators from the directory with caching
- * @returns Promise<Array<CreatorWithNFTData>> Array of creators with their NFT data and basename resolution
- */
-export async function getCreators(): Promise<CreatorWithNFTData[]> {
-  try {
-    // Get both cached data and TTL in one atomic operation
-    const multi = kv.multi();
-    multi.get(CREATORS_CACHE_KEY);
-    multi.ttl(CREATORS_CACHE_KEY);
-    const [cached, ttl] = (await multi.exec()) as [CachedData | null, number];
-
-    // Revalidate if TTL is negative (expired) or less than revalidation window
-    if (ttl < REVALIDATION_WINDOW) {
-      // Try to acquire the revalidation lock
-      const acquired = await kv.set(REVALIDATION_LOCK_KEY, Date.now(), {
-        nx: true,
-        ex: LOCK_DURATION,
-      });
-
-      // If we got the lock, trigger revalidation
-      if (acquired) {
-        if (cached?.data) {
-          // If we have cached data, revalidate in background
-          revalidateCreators().catch(console.error);
-        } else {
-          // If no cached data, revalidate immediately
-          return await refreshCreatorsData();
-        }
-      }
-    }
-
-    // Return cached data if we have it
-    if (cached?.data) {
-      return cached.data;
-    }
-
-    // If we somehow have no cached data (first ever run), fetch and cache
-    return await refreshCreatorsData();
-  } catch (error) {
-    console.error("Error fetching creators:", error);
-    // If there's an error but we have cached data, return it
-    const cached = await kv.get<CachedData>(CREATORS_CACHE_KEY);
-    if (cached?.data) {
-      return cached.data;
-    }
-    return [];
-  }
-}
-
-/**
- * Background revalidation function that updates the cache
- */
-export async function revalidateCreators(): Promise<void> {
-  try {
-    await refreshCreatorsData();
-  } catch (error) {
-    console.error("Error revalidating creators:", error);
-  } finally {
-    // Clean up the lock regardless of success/failure
-    await kv.del(REVALIDATION_LOCK_KEY);
-  }
-}
-
-/**
- * Utility function to process creators in chunks
- * @param items Array of items to process
- * @param chunkSize Size of each chunk
- * @param processor Function to process each chunk
- */
-async function processInChunks<T, R>(
-  items: T[],
-  chunkSize: number,
-  processor: (chunk: T[]) => Promise<R[]>,
-): Promise<R[]> {
-  const results: R[] = [];
-  for (let i = 0; i < items.length; i += chunkSize) {
-    const chunk = items.slice(i, i + chunkSize);
-    const chunkResults = await processor(chunk);
-    results.push(...chunkResults);
-  }
-  return results;
-}
-
-/**
- * Batch process creator data with retries and timeouts
- */
-async function batchProcessCreators(
-  creators: Creator[],
-  chunkSize = 20,
-): Promise<CreatorWithNFTData[]> {
-  return processInChunks(creators, chunkSize, async (chunk) => {
-    const processWithTimeout = async (
-      creator: Creator,
-    ): Promise<CreatorWithNFTData> => {
-      const timeout = 5000; // 5 second timeout
-      try {
-        const result = await Promise.race([
-          Promise.all([
-            getNFTContracts(creator.address).catch(() => []),
-            resolveBasenameOrAddress(creator.address).catch(() => null),
-            getGivesForCreator(creator.address).catch(() => []),
-          ]),
-          new Promise((_, reject) =>
-            setTimeout(() => reject(new Error("Timeout")), timeout),
-          ),
-        ]);
-
-        const [contracts, resolution, gives] = result as [any[], any, any[]];
-
-        const formattedResolution = resolution
-          ? {
-              basename: resolution.basename,
-              address: resolution.address,
-              resolved: !!resolution.basename,
-              textRecords: resolution.textRecords,
-            }
-          : null;
-
-        return {
-          ...creator,
-          resolution: formattedResolution,
-          gives: gives,
-          nftData: {
-            collections: contracts.map((contract) => ({
-              id: contract.contractAddress,
-              name: contract.name,
-              description: contract.description,
-              imageUrl: contract.imageUrl,
-              bannerImageUrl: contract.bannerImageUrl,
-              openseaUrl: `https://opensea.io/assets/base/${contract.contractAddress}`,
-              projectUrl: contract.projectUrl || "",
-              contractAddress: contract.contractAddress,
-            })),
-          },
-        };
-      } catch (error) {
-        console.error(`Error or timeout processing ${creator.address}:`, error);
-        return {
-          ...creator,
-          resolution: null,
-          gives: [],
-          nftData: { collections: [] },
-        };
-      }
-    };
-
-    return Promise.all(chunk.map(processWithTimeout));
-  });
-}
-
-async function getCreatorsFromAPI(): Promise<CreatorWithNFTData[]> {
-  const startTime = performance.now();
-  debugLog("Starting getCreatorsFromAPI");
-
-  const { data } = await getApolloClientAuthed().query({
-    query: gql`
-      query CreatorsDirGetAllCreators($circleId: bigint!) {
-        users(
-          where: {
-            circle_id: { _eq: $circleId }
-            profile: {
-              address: { _neq: "0x4fd59e958a4eaf440d761c41c73e40bffd069f4d" }
-            }
-          }
-          order_by: { created_at: desc }
-        ) {
-          id
-          profile {
-            id
-            address
-            name
-            avatar
-            description
-            farcaster_account {
-              username
-            }
-          }
-        }
-      }
-    `,
-    variables: {
-      circleId: CIRCLE_ID,
-    },
-  });
-
-  const queryTime = performance.now();
-  debugLog(
-    `GraphQL query completed in ${(queryTime - startTime).toFixed(2)}ms`,
-  );
-
-  // Transform the data to a more convenient format
-  const creators: Creator[] = data.users.map(
-    (user: {
-      id: string;
-      profile?: {
-        address?: string;
-        name?: string;
-        avatar?: string;
-        description?: string;
-        farcaster_account?: {
-          username?: string;
-        };
-      };
-    }) => ({
-      id: user.id,
-      address: user.profile?.address || "",
-      name: user.profile?.name || "",
-      avatar: user.profile?.avatar
-        ? user.profile.avatar.startsWith("http")
-          ? user.profile.avatar
-          : `https://coordinape-prod.s3.amazonaws.com/${user.profile.avatar}`
-        : "",
-      description: user.profile?.description || "",
-      farcasterUsername: user.profile?.farcaster_account?.username || "",
-    }),
-  );
-
-  const transformTime = performance.now();
-  debugLog(
-    `Data transformation completed in ${(transformTime - queryTime).toFixed(2)}ms`,
-  );
-  debugLog(
-    `Processing ${creators.length} creators for NFT data and basenames...`,
-  );
-
-  // Process creators in batches with improved error handling and timeouts
-  const creatorsWithNFTData = await batchProcessCreators(creators);
-
-  const endTime = performance.now();
-  const totalTime = endTime - startTime;
-  console.log(`getCreatorsFromAPI completed in ${totalTime.toFixed(2)}ms`);
-  debugLog(
-    `Average time per creator: ${(totalTime / creators.length).toFixed(2)}ms`,
-  );
-
-  return creatorsWithNFTData;
-}
-
-const groupedGives = (gives: Give[]): GroupedGives => {
-  return gives.reduce((acc, give) => {
-    if (give.skill) {
-      if (!acc[give.skill]) {
-        acc[give.skill] = [];
-      }
-      acc[give.skill].push(give);
-    } else {
-      if (!acc[""]) {
-        acc[""] = [];
-      }
-      acc[""].push(give);
-    }
-    return acc;
-  }, {} as GroupedGives);
-};
-
-const sortedGives = (groupedGives: GroupedGives): SortedGiveGroup[] => {
-  return Object.entries(groupedGives)
-    .map(([skill, gives]) => ({
-      count: gives.length,
-      gives: gives,
-      skill: skill,
-    }))
-    .sort((a, b) => {
-      const diff = b.count - a.count;
-      if (diff === 0) {
-        return a.skill.localeCompare(b.skill);
-      }
-      return diff;
-    });
-};
-
-const groupAndSortGives = (gives: Give[]): SortedGiveGroup[] => {
-  return sortedGives(groupedGives(gives));
-};
-
-/**
- * Fetches all gives for a specific creator address
- * @param address Ethereum address of the creator
- * @returns Promise<SortedGiveGroup[]> Array of sorted and grouped gives for the creator
- */
-export async function getGivesForCreator(
-  address: string,
-): Promise<SortedGiveGroup[]> {
-  try {
-    const { data } = await getApolloClient().query({
-      query: gql`
-        query CreatorsDirGetCreatorGives($address: String!) {
-          colinks_gives(
-            where: {
-              target_profile_public: { address: { _ilike: $address } }
-              skill: { _is_null: false }
-            }
-          ) {
-            id
-            skill
-            created_at
-          }
-        }
-      `,
-      variables: {
-        address: address,
-      },
-    });
-
-    // console.log("data for gives", JSON.stringify(data, null, 2));
-    return groupAndSortGives(data.colinks_gives);
-  } catch (error) {
-    console.error(`Error fetching gives for address ${address}:`, error);
-    return [];
   }
 }
